moved plugin related stuff to plugin_manager
